Fixed warnings on MSVC 6.0.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/163



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@476 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/compitr.cpp
@@ -99,7 +99,7 @@ void ComposedCharIter::getDecomposition(UnicodeString& result) const
         Normalizer::doAppend(DecompData::contents, pos, result);
     } 
     else if (hangul && curChar >= Normalizer::HANGUL_BASE && curChar < Normalizer::HANGUL_LIMIT) {
-        Normalizer::hangulToJamo(curChar, result, minDecomp);
+        Normalizer::hangulToJamo(curChar, result, (uint16_t)minDecomp);
     } 
     else {
         result += curChar;
