+++ source/test/intltest/tsdate.cpp
@@ -222,8 +222,8 @@ double IntlTestDateFormat::randDouble()
     if (d > 0.0)
     {
         double e = uprv_floor(uprv_log10(d));
-        if (e < -2.0) d *= uprv_pow10(-e-2);
-        else if (e > -1.0) d /= uprv_pow10(e+1);
+        if (e < -2.0) d *= uprv_pow10((long)-e-2);
+        else if (e > -1.0) d /= uprv_pow10((long)e+1);
     }
     return d;
 }
